api-python/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.14
langchain-mistralai==0.2.4
python-dotenv==1.0.0
pydantic==2.10.4
kagglehub==0.3.6
